Move common controller logic into util
