+++ cmd/feed-dns/main.go
@@ -3,21 +3,11 @@ package main
 import (
 	"flag"
 	"os"
-	"os/signal"
-
-	"io/ioutil"
-	"net/http"
-	_ "net/http/pprof"
-
-	"strconv"
-
-	"io"
-
-	"syscall"
 
 	log "github.com/Sirupsen/logrus"
 	"github.com/sky-uk/feed/dns"
 	"github.com/sky-uk/feed/k8s"
+	"github.com/sky-uk/feed/util"
 )
 
 var (
@@ -45,12 +35,13 @@ func init() {
 
 func main() {
 	flag.Parse()
-	configureLogging()
+	util.ConfigureLogging(debug)
 
-	controller := dns.New(createK8sClient())
+	client := k8s.New(caCertFile, tokenFile, apiServer)
+	controller := dns.New(client)
 
-	configureHealthPort(controller)
-	addSignalHandler(controller)
+	util.ConfigureHealthPort(controller, healthPort)
+	util.AddSignalHandler(controller)
 
 	err := controller.Start()
 	if err != nil {
@@ -59,75 +50,4 @@ func main() {
 	}
 
 	select {}
-}
-
-func configureLogging() {
-	// logging is the main output, so write it all to stdout
-	log.SetOutput(os.Stdout)
-	if debug {
-		log.SetLevel(log.DebugLevel)
-	}
-}
-
-func createK8sClient() k8s.Client {
-	caCert := readFile(caCertFile)
-	token := string(readFile(tokenFile))
-
-	client, err := k8s.New(apiServer, caCert, token)
-	if err != nil {
-		log.Errorf("Unable to create Kubernetes client: %v", err)
-		os.Exit(-1)
-	}
-
-	return client
-}
-
-func readFile(path string) []byte {
-	data, err := ioutil.ReadFile(path)
-	if err != nil {
-		log.Errorf("Unable to read %s: %v", path, err)
-		os.Exit(-1)
-	}
-	return data
-}
-
-func configureHealthPort(controller dns.Controller) {
-	http.HandleFunc("/health", checkHealth(controller))
-
-	go func() {
-		log.Error(http.ListenAndServe(":"+strconv.Itoa(healthPort), nil))
-		log.Info(controller.Stop())
-		os.Exit(-1)
-	}()
-}
-
-func checkHealth(controller dns.Controller) func(w http.ResponseWriter, r *http.Request) {
-	return func(w http.ResponseWriter, r *http.Request) {
-		if !controller.Healthy() {
-			w.WriteHeader(http.StatusServiceUnavailable)
-			io.WriteString(w, "fail\n")
-			return
-		}
-
-		w.WriteHeader(http.StatusOK)
-		io.WriteString(w, "ok\n")
-	}
-}
-
-func addSignalHandler(controller dns.Controller) {
-	c := make(chan os.Signal, 1)
-	// SIGTERM is used by Kubernetes to gracefully stop pods.
-	signal.Notify(c, syscall.SIGINT, syscall.SIGTERM)
-
-	go func() {
-		for sig := range c {
-			log.Infof("Signalled %v, shutting down gracefully", sig)
-			err := controller.Stop()
-			if err != nil {
-				log.Error("Error while stopping controller: ", err)
-				os.Exit(-1)
-			}
-			os.Exit(0)
-		}
-	}()
 }
